requirements.txt
Django>=3.0,<4.0
Pillow==8.3.2
django-anymail==8.4
django-axes==5.26.0
django-cryptography==1.0
django-crispy-forms==1.13.0
django-csp==3.7
django-datatable-view==2.1.6
django-debug-toolbar==3.2.2
django-formtools==2.3
django-multiselectfield==0.1.12
django-simple-captcha==0.5.14
django-resized==0.3.11
django-select2==7.7.3
django-sn==0.8.11.9
django-extensions==3.1.3
django-upload-form==0.5.0
django-dbbackup==3.3.0
django-storages==1.13.1
django-crontab==0.7.1
django-cleanup==7.0.0
dropbox==11.23.0
numpy==1.21.2
opencv-python==4.5.3.56
pycryptodome==3.10.4
python-dateutil==2.8.2
python-gnupg==0.4.7
reportlab==3.6.1
bleach==4.1.0
mysqlclient==2.0.3
jellyfish==0.8.9
lxml==4.9.0
eopayment==1.79
colorhash==1.2.1
feedparser==6.0.10
beautifulsoup4==4.11.1
pystrich==0.8
docx-mailmerge2